// teeok-burak/src/lib/enums/product.enum.ts
export enum ProductCollection {
  GARAETTEOK = "GARAETTEOK",
  SONGPYEON = "SONGPYEON",
  INJEOLMI = "INJEOLMI",
  TTEOKBOKKI = "TTEOKBOKKI",
  GIFT_SET = "GIFT_SET",
  OTHER = "OTHER",
}

export enum ProductSize {
  SMALL = "SMALL",
  MEDIUM = "MEDIUM",
  LARGE = "LARGE",
  SET = "SET",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}